[refactor] CategoriesList - extract query params to hook
